# gest-ashoex-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Health Check",
        "description": "API para verificar el estado de salud del sistema",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/grupo/{id}": {
            "get": {
                "tags": [
                    "Grupos"
                ],
                "summary": "Obtiene los detalles de una grupo por su ID",
                "operationId": "71ecd5bbf3a284f66ce9def4a875954c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la grupo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "materia_id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "nro_grupo": {
                                                    "type": "number",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": []
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Operación exitosa"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "El grupo  no existe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {},
                                            "example": []
                                        },
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Grupo no encontrado"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Operación fallida"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error en el servidor interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {},
                                            "example": []
                                        },
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Ocurrió un error inesperado"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Operación fallida"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}